refactor: install scss, split styles, change file structure
