+++ src/components/Field/Field.js
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import Cell from '../Cell/Cell';
 import SIZES from '../../constants/SIZES';
 import MINES from '../../constants/MINES';
-import createEmptyField from './createEmptyField';
-import fillField from './fillField';
-import revealEmptyCells from '../Cell/revealEmptyCells';
+import createEmptyField from './utils/createEmptyField';
+import fillField from './utils/fillField';
+import revealEmptyCells from '../Cell/utils/revealEmptyCells';
 import { revealWholeField, playGameLost } from '../../utils/gameLost';
 import gameWon from '../../utils/gameWon';
 import useWindowWidth from '../../utils/useWindowWidth';
