+++ src/components/Field/utils/createEmptyField.js
@@ -1,4 +1,4 @@
-import SIZES from '../../constants/SIZES';
+import SIZES from '../../../constants/SIZES';
 
 export default function createEmptyField() {
   const grid = Array(SIZES.fieldArea).fill(null).map((item, index) => ({
