+++ src/styles/modules/common.scss
@@ -1,0 +1,15 @@
+* {
+  margin: 0;
+  font-family: Helvetica, Arial, Sans-Serif;
+}
+
+body {
+  background-color: #30353b;
+}
+
+.field {
+  margin: auto;
+  padding: 10vh 20px;
+  display: flex;
+  flex-wrap: wrap;
+}
