+++ src/styles/styles.scss
@@ -1,102 +1,3 @@
-$depth: 4;
-$color-prim: #5ba6ff;
-$color-sec: #f3d248;
-$lighten-on-hover: 6%;
-$button-speed: .06;
-$timing-function: ease-out;
-
-@mixin background($imageName, $color) {
-  background-image: url('../assets/images/#{$imageName}.png');
-  background-size: contain;
-  background-color: $color;
-}
-
-* {
-  margin: 0;
-  font-family: Helvetica, Arial, Sans-Serif;
-}
-
-body {
-  background-color: #30353b;
-}
-
-.field {
-  margin: auto;
-  padding: 10vh 20px;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.cell {
-  position: relative;
-  top: 0;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  cursor: default;
-  text-decoration: none;
-  font-weight: bold;
-
-  background: linear-gradient(to bottom,
-      saturate(lighten($color-prim, 15%), 35%) 0%,
-      saturate(lighten($color-prim, 12%), 15%) 26%, $color-prim 100%);
-  border-radius: 4px;
-  border-top: 1px solid lighten($color-prim, 22%);
-  border-bottom: 1px solid lighten($color-prim, 20%);
-  transition: all ($button-speed + s) $timing-function;
-}
-
-.cell-default:hover {
-  background: linear-gradient(to bottom,
-      lighten(saturate(lighten($color-prim, 15%), 35%), $lighten-on-hover) 0%,
-      lighten(saturate(lighten($color-prim, 12%), 15%), $lighten-on-hover) 26%,
-      lighten($color-prim, $lighten-on-hover) 100%);
-}
-
-.cell-default:active {
-  top: $depth+px;
-
-  &:before {
-    top: 0;
-    box-shadow: 0 3px 3px rgba(0, 0, 0, 1), 0 3px 9px rgba(0, 0, 0, .2);
-  }
-}
-
-.cell:before {
-  content: '';
-  position: absolute;
-  top: $depth+px;
-  right: 0;
-  left: 0;
-  z-index: -1;
-
-  display: inline-block;
-  border-radius: 5px;
-  height: 100%;
-  background: linear-gradient(to top,
-      saturate(darken($color-prim, 40%), 15%) 0%,
-      saturate(darken($color-prim, 25%), 10%) $depth+px);
-  transition: all (($button-speed*1.3) + s) $timing-function;
-  box-shadow:
-    0 1px 0 1px rgba(0, 0, 0, .3),
-    0 2px ($depth*.2+px) rgba(0, 0, 0, .5),
-    0 ($depth*1+px) 9px rgba(0, 0, 0, .2);
-}
-
-.cell-default {
-  cursor: pointer;
-}
-
-.cell-flagged {
-  @include background('flag', $color-sec);
-}
-
-.cell-revealed {
-  background: #D8D8D8;
-}
-
-.cell-mine {
-  @include background('mine', $color-prim);
-}
+@import './modules/constants';
+@import './modules/common';
+@import './modules/cell';
